docker/predict/requirements.txt
joblib
pandas==2.2.2
numpy==2.0.2
pyarrow==16.1.0
google-cloud-bigquery
gcsfs